Small updates
- Fixed environment that could provide infeasible instances for dynamic variant where it was not possible to return to the depot in time. These requests will now be filtered or be indicated as must-go before they become infeasible. It was tested that this does not affect the qualification instances.
- Print HGS cost as integer (sometimes a large cost was printed in scientific notation causing problems)
- Updated Slack URLs and added text about leaderboard prizes

+++ baselines/hgs_vrptw/Individual.cpp
@@ -177,7 +177,7 @@ void Individual::exportCVRPLibFormat(std::string fileName)
 				myfile << std::endl;
 			}
 		}
-		myfile << "Cost " << myCostSol.penalizedCost << std::endl;
+		myfile << "Cost " << (int) myCostSol.penalizedCost << std::endl;
 		myfile << "Time " << params->getTimeElapsedSeconds() << std::endl;
 	}
 	else std::cout << "----- IMPOSSIBLE TO OPEN: " << fileName << std::endl;
@@ -198,7 +198,7 @@ void Individual::printCVRPLibFormat()
 			std::cout << std::endl;
 		}
 	}
-	std::cout << "Cost " << myCostSol.penalizedCost << std::endl;
+	std::cout << "Cost " << (int) myCostSol.penalizedCost << std::endl;
 	std::cout << "Time " << params->getTimeElapsedSeconds() << std::endl;
 	fflush(stdout);
 }
